+++ nemo/collections/nlp/modules/common/token_classifier.py
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from dataclasses import MISSING, dataclass
 from typing import Dict, Optional
 
 from torch import nn as nn
@@ -24,6 +25,18 @@ from nemo.core.neural_types import LogitsType, LogprobsType, NeuralType
 __all__ = ['BertPretrainingTokenClassifier', 'TokenClassifier']
 
 ACT2FN = {"gelu": nn.functional.gelu, "relu": nn.functional.relu}
+
+
+@dataclass
+class TokenClassifierConfig:
+    hidden_size: int = MISSING
+    num_classes: int = MISSING
+    num_layers: int = 1
+    activation: str = 'relu'
+    log_softmax: bool = True
+    dropout: float = 0.0
+    use_transformer_init: bool = True
+    _target_: str = 'nemo.collections.nlp.modules.common.token_classifier.TokenClassifier'
 
 
 class TokenClassifier(Classifier):
