+++ nemo/collections/nlp/modules/common/transformer/transformer_modules.py
@@ -14,6 +14,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from dataclasses import MISSING, dataclass
 import math
 
 import torch
@@ -47,6 +48,17 @@ class FixedPositionalEncoding(nn.Module):
 
     def forward(self, position_ids):
         return torch.embedding(self.pos_enc, position_ids)
+
+
+@dataclass
+class TransformerEmbeddingConfig:
+    vocab_size: int = MISSING
+    hidden_size: int = MISSING
+    max_sequence_length: int = 512
+    num_token_types: int = 2
+    embedding_dropout: float = 0.0
+    learn_positional_encodings: bool = False
+    _target_: str = "nemo.collections.nlp.modules.common.transformer.TransformerEmbedding"
 
 
 class TransformerEmbedding(nn.Module):
